v0.20 完成分片上传基本功能，并通过简单的集成测试

+++ server/partInfoStorage.go
@@ -22,6 +22,6 @@ type CompleteExtra struct {
 
 type UploadPartInfo struct {
 	Etag       string
-	PartNumber int64
-	PartSize   int
+	PartNumber int
+	PartSize   int64
 }
